requirements.txt
selenium==4.21.0
pytest==7.4.4
allure-pytest==2.13.2
webdriver-manager==4.0.1